+++ src/sisl/io/siesta/tests/test_vibra.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 
 import os.path as osp
 
+import numpy as np
 import pytest
 
 import sisl
@@ -26,15 +27,31 @@ def test_eigenmode_read(sisl_files):
     nlines = len(blines)
     iklast = [sum(int(l.split()[0]) for l in blines[: i + 1]) for i in range(nlines)]
     klast = [list(map(float, l.split()[1:4])) for l in blines]
-    print(iklast, klast)
 
     # yield modes
     nmodes = 0
     nmodes_total = 0
+    ks = []
     for state in vectors.yield_eigenmode():
         nmodes += 1
         nmodes_total += len(state)
+        ks.append(state.info["k"])
 
-    print(nmodes, nmodes_total)
     assert nmodes == nk
     assert nmodes_total == geometry.na * 3 * nk
+
+    bz = vectors.read_brillouinzone()
+    assert len(bz) == nk
+    assert np.allclose(bz.k, ks)
+
+
+def test_eigenmode_values(sisl_files):
+    fdf = sisl.get_sile(sisl_files(_dir, "h_chain_vibra.fdf"))
+    vectors = sisl.io.siesta.vectorsSileSiesta(
+        sisl_files(_dir, "h_chain_vibra.vectors"),
+        geometry=fdf.read_geometry(),
+    )
+
+    mode = vectors.read_eigenmode()
+    assert np.allclose(mode.state[0, 0:3], [-0.7071, -0.2400e-4, -0.2400e-4])
+    assert np.allclose(mode.state[0, 3:7], [0.7071, 0.2400e-4, 0.2400e-4])